requirements.txt
click==6.7
decorator==4.3.0
entrypoint2==0.0.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.2.0
python-sane==2.8.2
Werkzeug==0.14.1
